cloud-functions/csv-receiver/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
flask==3.*
werkzeug==3.*
